feat: enhance entity interpolation by refining coordinate packing and unpacking

+++ server/src/packet/PacketFactory/UDP/UDPPacketFactory.cpp
@@ -120,8 +120,8 @@ namespace Net::Factory
 
                         SnapshotEntityData packed{};
                         packed.id = htonl(static_cast<uint32_t>(id));
-                        packed.x = htons(static_cast<uint16_t>(x));
-                        packed.y = htons(static_cast<uint16_t>(y));
+                        packed.x = htons(static_cast<uint16_t>(static_cast<int16_t>(x)));
+                        packed.y = htons(static_cast<uint16_t>(static_cast<int16_t>(y)));
                         packed.spriteId = static_cast<uint8_t>(spriteId);
 
                         std::memcpy(rawBuf.data() + off, &packed, sizeof(packed));
